# apps/backend/app/modules/pak/router.py
from typing import Annotated, Literal, cast
from uuid import UUID

from fastapi import APIRouter, Depends, Query, Request, status

from app.api.auth_deps import CurrentPrincipalDep, require_permission
from app.modules.defects.schemas import DefectGroupSummaryResponse

from .enums import PakDeviceKind, PakStatus
from .exceptions import PakNotFoundError, PakTestNotFoundError
from .models import PakDevice, PakTest
from .permissions import PakPermission
from .schemas import (
    CreatePakDeviceRequest,
    CreatePakDeviceResponse,
    PakAccessKeyResponse,
    PakDeviceListResponse,
    PakDeviceResponse,
    PakTestListResponse,
    PakTestResponse,
    UpdateActiveRequest,
    UpdateArchivedRequest,
    UpdatePakDeviceRequest,
)
from .services import PakManagementService, PakTestCatalogService

router = APIRouter(prefix="/pak", tags=["pak"])


def _response(pak: PakDevice) -> PakDeviceResponse:
    return PakDeviceResponse(
        id=pak.id,
        code=pak.code,
        kind=pak.kind,
        oauth_client_id=pak.oauth_client_id,
        status=PakStatus.ACTIVE if pak.is_active else PakStatus.INACTIVE,
        last_seen_at=pak.last_seen_at,
        archived_at=pak.archived_at,
    )


def _service(request: Request) -> PakManagementService:
    return cast(PakManagementService, request.app.state.pak_management)


def _test_response(test: PakTest) -> PakTestResponse:
    return PakTestResponse(
        id=test.id,
        test_name=test.test_name,
        test_label=test.test_label,
        defect_group_id=test.defect_group_id,
        defect_group=DefectGroupSummaryResponse(
            id=test.defect_group.id,
            code=test.defect_group.code,
            name=test.defect_group.name,
            archived_at=test.defect_group.archived_at,
        ),
        last_seen_at=test.last_seen_at,
        created_at=test.created_at,
        updated_at=test.updated_at,
    )


def _test_service(request: Request) -> PakTestCatalogService:
    return cast(PakTestCatalogService, request.app.state.pak_test_catalog)


@router.get("", response_model=PakDeviceListResponse)
async def list_pak(
    _: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.READ)),
    ],
    request: Request,
    q: str | None = None,
    kind: PakDeviceKind | None = None,
    status: PakStatus | None = None,
    archived: bool = False,
    page: int = Query(default=1, ge=1),
    page_size: int = Query(default=25, ge=1, le=100),
    sort: Literal[
        "code",
        "kind",
        "created_at",
        "last_seen_at",
        "archived_at",
    ] = "code",
    order: Literal["asc", "desc"] = "asc",
) -> PakDeviceListResponse:
    paks, total = await _service(request).list(
        q=q,
        kind=kind,
        active=None if status is None else status is PakStatus.ACTIVE,
        archived=archived,
        page=page,
        page_size=page_size,
        sort=sort,
        order=order,
    )

    return PakDeviceListResponse(
        items=[_response(pak) for pak in paks],
        total=total,
        page=page,
        page_size=page_size,
    )


@router.get("/tests", response_model=PakTestListResponse)
async def list_pak_tests(
    _: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.READ)),
    ],
    request: Request,
    q: str | None = None,
    defect_group_id: UUID | None = None,
    page: int = Query(default=1, ge=1),
    page_size: int = Query(default=25, ge=1, le=100),
    sort: Literal[
        "test_name",
        "test_label",
        "last_seen_at",
        "created_at",
        "updated_at",
    ] = "test_name",
    order: Literal["asc", "desc"] = "asc",
) -> PakTestListResponse:
    tests, total = await _test_service(request).list(
        q=q,
        defect_group_id=defect_group_id,
        page=page,
        page_size=page_size,
        sort=sort,
        order=order,
    )

    return PakTestListResponse(
        items=[_test_response(test) for test in tests],
        total=total,
        page=page,
        page_size=page_size,
    )


@router.get("/tests/{test_id}", response_model=PakTestResponse)
async def get_pak_test(
    test_id: UUID,
    _: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.READ)),
    ],
    request: Request,
) -> PakTestResponse:
    test = await _test_service(request).get(test_id)

    if test is None:
        raise PakTestNotFoundError

    return _test_response(test)


@router.get("/{pak_id}", response_model=PakDeviceResponse)
async def get_pak(
    pak_id: UUID,
    _: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.READ)),
    ],
    request: Request,
) -> PakDeviceResponse:
    pak = await _service(request).get(pak_id)

    if pak is None:
        raise PakNotFoundError

    return _response(pak)


@router.post("", response_model=CreatePakDeviceResponse, status_code=status.HTTP_201_CREATED)
async def create_pak(
    payload: CreatePakDeviceRequest,
    principal: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.CREATE)),
    ],
    request: Request,
) -> CreatePakDeviceResponse:
    pak, access_key = await _service(request).create(
        actor=principal,
        code=payload.code,
        kind=payload.kind,
        active=payload.active,
    )

    return CreatePakDeviceResponse(device=_response(pak), access_key=access_key)


@router.get("/{pak_id}/access-key", response_model=PakAccessKeyResponse)
async def get_access_key(
    pak_id: UUID,
    principal: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.READ_ACCESS_KEY)),
    ],
    request: Request,
) -> PakAccessKeyResponse:
    access_key = await _service(request).get_access_key(
        actor=principal,
        pak_id=pak_id,
    )

    return PakAccessKeyResponse(access_key=access_key)


@router.post("/{pak_id}/access-key/rotate", response_model=PakAccessKeyResponse)
async def rotate_access_key(
    pak_id: UUID,
    principal: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.ROTATE_ACCESS_KEY)),
    ],
    request: Request,
) -> PakAccessKeyResponse:
    access_key = await _service(request).rotate_access_key(
        actor=principal,
        pak_id=pak_id,
    )

    return PakAccessKeyResponse(access_key=access_key)


@router.patch("/{pak_id}", response_model=PakDeviceResponse)
async def update_pak(
    pak_id: UUID,
    payload: UpdatePakDeviceRequest,
    principal: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.UPDATE)),
    ],
    request: Request,
) -> PakDeviceResponse:
    return _response(
        await _service(request).update(
            actor=principal,
            pak_id=pak_id,
            code=payload.code,
            kind=payload.kind,
        )
    )


@router.put("/{pak_id}/active", response_model=PakDeviceResponse)
async def update_active(
    pak_id: UUID,
    payload: UpdateActiveRequest,
    principal: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.SET_ACTIVE)),
    ],
    request: Request,
) -> PakDeviceResponse:
    return _response(
        await _service(request).set_active(
            actor=principal,
            pak_id=pak_id,
            active=payload.active,
        )
    )


@router.put("/{pak_id}/archived", response_model=PakDeviceResponse)
async def update_archived(
    pak_id: UUID,
    payload: UpdateArchivedRequest,
    principal: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.ARCHIVE)),
    ],
    request: Request,
) -> PakDeviceResponse:
    return _response(
        await _service(request).set_archived(
            actor=principal,
            pak_id=pak_id,
            archived=payload.archived,
        )
    )


@router.delete("/{pak_id}", status_code=status.HTTP_204_NO_CONTENT)
async def delete_pak(
    pak_id: UUID,
    principal: Annotated[
        CurrentPrincipalDep,
        Depends(require_permission(PakPermission.DELETE)),
    ],
    request: Request,
) -> None:
    await _service(request).delete(
        actor=principal,
        pak_id=pak_id,
    )