# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatePakDeviceRequest": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "code": {
            "maxLength": 255,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Code",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/PakDeviceKind"
          }
        },
        "required": [
          "code",
          "kind"
        ],
        "title": "CreatePakDeviceRequest",
        "type": "object"
      },
      "CreatePakDeviceResponse": {
        "properties": {
          "access_key": {
            "title": "Access Key",
            "type": "string"
          },
          "device": {
            "$ref": "#/components/schemas/PakDeviceResponse"
          }
        },
        "required": [
          "device",
          "access_key"
        ],
        "title": "CreatePakDeviceResponse",
        "type": "object"
      },
...
      "DefectGroupSummaryResponse": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "archived_at"
        ],
        "title": "DefectGroupSummaryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PakAccessKeyResponse": {
        "properties": {
          "access_key": {
            "title": "Access Key",
            "type": "string"
          }
        },
        "required": [
          "access_key"
        ],
        "title": "PakAccessKeyResponse",
        "type": "object"
      },
      "PakDeviceKind": {
        "enum": [
          "engineering",
          "otk_line"
        ],
        "title": "PakDeviceKind",
        "type": "string"
      },
      "PakDeviceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PakDeviceResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PakDeviceListResponse",
        "type": "object"
      },
      "PakDeviceResponse": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/PakDeviceKind"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "oauth_client_id": {
            "title": "Oauth Client Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PakStatus"
          }
        },
        "required": [
          "id",
          "code",
          "kind",
          "oauth_client_id",
          "status",
          "last_seen_at",
          "archived_at"
        ],
        "title": "PakDeviceResponse",
        "type": "object"
      },
...
      "PakStatus": {
        "enum": [
          "active",
          "inactive"
        ],
        "title": "PakStatus",
        "type": "string"
      },
      "PakTestListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PakTestResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PakTestListResponse",
        "type": "object"
      },
      "PakTestResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "defect_group": {
            "$ref": "#/components/schemas/DefectGroupSummaryResponse"
          },
          "defect_group_id": {
            "format": "uuid",
            "title": "Defect Group Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "test_label": {
            "title": "Test Label",
            "type": "string"
          },
          "test_name": {
            "title": "Test Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "test_name",
          "test_label",
          "defect_group_id",
          "defect_group",
          "last_seen_at",
          "created_at",
          "updated_at"
        ],
        "title": "PakTestResponse",
        "type": "object"
      },
...
      "UpdateActiveRequest": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          }
        },
        "required": [
          "active"
        ],
        "title": "UpdateActiveRequest",
        "type": "object"
      },
      "UpdateArchivedRequest": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          }
        },
        "required": [
          "archived"
        ],
        "title": "UpdateArchivedRequest",
        "type": "object"
      },
...
      "UpdatePakDeviceRequest": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PakDeviceKind"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "UpdatePakDeviceRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "backend",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/pak": {
      "get": {
        "operationId": "pak-list_pak",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PakDeviceKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PakStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "code",
              "enum": [
                "code",
                "kind",
                "created_at",
                "last_seen_at",
                "archived_at"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PakDeviceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pak",
        "tags": [
          "pak"
        ]
      },
      "post": {
        "operationId": "pak-create_pak",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePakDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePakDeviceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Pak",
        "tags": [
          "pak"
        ]
      }
    },
    "/pak/tests": {
      "get": {
        "operationId": "pak-list_pak_tests",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "defect_group_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Defect Group Id"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "test_name",
              "enum": [
                "test_name",
                "test_label",
                "last_seen_at",
                "created_at",
                "updated_at"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PakTestListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pak Tests",
        "tags": [
          "pak"
        ]
      }
    },
    "/pak/tests/{test_id}": {
      "get": {
        "operationId": "pak-get_pak_test",
        "parameters": [
          {
            "in": "path",
            "name": "test_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Test Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PakTestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pak Test",
        "tags": [
          "pak"
        ]
      }
    },
    "/pak/{pak_id}": {
      "delete": {
        "operationId": "pak-delete_pak",
        "parameters": [
          {
            "in": "path",
            "name": "pak_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pak Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Pak",
        "tags": [
          "pak"
        ]
      },
      "get": {
        "operationId": "pak-get_pak",
        "parameters": [
          {
            "in": "path",
            "name": "pak_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pak Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PakDeviceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pak",
        "tags": [
          "pak"
        ]
      },
      "patch": {
        "operationId": "pak-update_pak",
        "parameters": [
          {
            "in": "path",
            "name": "pak_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pak Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePakDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PakDeviceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Pak",
        "tags": [
          "pak"
        ]
      }
    },
    "/pak/{pak_id}/access-key": {
      "get": {
        "operationId": "pak-get_access_key",
        "parameters": [
          {
            "in": "path",
            "name": "pak_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pak Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PakAccessKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Access Key",
        "tags": [
          "pak"
        ]
      }
    },
    "/pak/{pak_id}/access-key/rotate": {
      "post": {
        "operationId": "pak-rotate_access_key",
        "parameters": [
          {
            "in": "path",
            "name": "pak_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pak Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PakAccessKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotate Access Key",
        "tags": [
          "pak"
        ]
      }
    },
    "/pak/{pak_id}/active": {
      "put": {
        "operationId": "pak-update_active",
        "parameters": [
          {
            "in": "path",
            "name": "pak_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pak Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PakDeviceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Active",
        "tags": [
          "pak"
        ]
      }
    },
    "/pak/{pak_id}/archived": {
      "put": {
        "operationId": "pak-update_archived",
        "parameters": [
          {
            "in": "path",
            "name": "pak_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pak Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArchivedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PakDeviceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Archived",
        "tags": [
          "pak"
        ]
      }
    },
...
  }
}